requirements.txt
Django==5.1
psycopg[binary]~=3.1
google-cloud-pubsub
protobuf~=5.28.0
django-environ~=0.11.2
debugpy>=1.8.0
djangorestframework==3.14.0